refactor the setting modal
