+++ windwatts-ui/src/services/api.ts
@@ -1,4 +1,3 @@
-import { get } from "http";
 import {
   EnergyProductionRequest,
   WindspeedByLatLngRequest,
